makes target reports polymorphic with other aggregate reports again. applies some code style feedback
