+++ aggregate-service/src/main/java/org/opentestsystem/rdw/olap/service/impl/AbstractAggregateQueryHandler.java
@@ -8,7 +8,6 @@ import org.opentestsystem.rdw.olap.repository.ReportRepository;
 import org.opentestsystem.rdw.olap.service.AggregateQueryHandler;
 import org.opentestsystem.rdw.olap.service.AuthorizationService;
 import org.opentestsystem.rdw.reporting.common.model.AbstractAggregateReportQuery;
-import org.opentestsystem.rdw.reporting.common.model.AbstractSubgroupedAggregateReportQuery;
 import org.opentestsystem.rdw.reporting.common.model.AggregateRow;
 import org.opentestsystem.rdw.reporting.common.model.DimensionType;
 import org.opentestsystem.rdw.reporting.common.model.StudentFilters;
@@ -103,22 +102,16 @@ abstract class AbstractAggregateQueryHandler<
      */
     protected Q preProcess(final Q queryIn) {
         Q query = queryIn.isFilteredSubgroupQuery()
-                ? (Q) ensureOverallSubgroup((AbstractSubgroupedAggregateReportQuery) queryIn)
+                ? ensureOverallSubgroup(queryIn)
                 : ensureDimension(queryIn, Overall);
 
         checkArgument(!query.getSubjectCodes().isEmpty(), "report query is missing subjects");
         checkArgument(!query.getAssessmentGradeCodes().isEmpty(), "report query is missing asmt grades");
         checkArgument(!query.getSchoolYears().isEmpty(), "report query is missing school years");
-        if (query instanceof AbstractSubgroupedAggregateReportQuery) {
-
-            final AbstractSubgroupedAggregateReportQuery subgroupedQuery = (AbstractSubgroupedAggregateReportQuery) query;
-
-            checkArgument(!(!subgroupedQuery.isIncludeState() && !subgroupedQuery.isIncludeAllDistricts() && query.getDistrictIds().isEmpty() && query.getSchoolIds().isEmpty()), "report query is missing organization");
-            if (subgroupedQuery.isIncludeState()) {
-                checkArgument(stateAggregateAssessmentTypes.contains(query.getAssessmentTypeCode()), "report query cannot include state results for " + query.getAssessmentTypeCode() + " assessment types");
-            }
+        checkArgument(!(!query.isIncludeState() && !query.isIncludeAllDistricts() && query.getDistrictIds().isEmpty() && query.getSchoolIds().isEmpty()), "report query is missing organization");
+        if (query.isIncludeState()) {
+            checkArgument(stateAggregateAssessmentTypes.contains(query.getAssessmentTypeCode()), "report query cannot include state results for " + query.getAssessmentTypeCode() + " assessment types");
         }
-
         return query;
     }
 
@@ -128,10 +121,10 @@ abstract class AbstractAggregateQueryHandler<
                 : (Q) (query.<AbstractAggregateReportQuery.Builder>copy()).dimensionType(Overall).build();
     }
 
-    private AbstractSubgroupedAggregateReportQuery ensureOverallSubgroup(final AbstractSubgroupedAggregateReportQuery query) {
+    private Q ensureOverallSubgroup(final Q query) {
         return query.getSubgroups().keySet().contains(Overall.name())
                 ? query
-                : (AbstractSubgroupedAggregateReportQuery) query.<AbstractSubgroupedAggregateReportQuery.Builder>copy().subgroup(Overall.name(), StudentFilters.builder().build()).build();
+                : (Q) (query.<AbstractAggregateReportQuery.Builder>copy()).subgroup(Overall.name(), StudentFilters.builder().build()).build();
     }
 
     private class RunQueryTask {
