+++ aggregate-service/src/main/java/org/opentestsystem/rdw/olap/service/impl/AbstractQueryConverter.java
@@ -5,7 +5,6 @@ import org.opentestsystem.rdw.olap.model.OrganizationTypeQuery;
 import org.opentestsystem.rdw.olap.service.AggregateQueryConverter;
 import org.opentestsystem.rdw.olap.sqlbuilder.impl.QueryProviderRepositoryHelper;
 import org.opentestsystem.rdw.reporting.common.model.AbstractAggregateReportQuery;
-import org.opentestsystem.rdw.reporting.common.model.AbstractSubgroupedAggregateReportQuery;
 import org.opentestsystem.rdw.reporting.common.model.DimensionType;
 import org.opentestsystem.rdw.reporting.common.model.OrganizationType;
 import org.opentestsystem.rdw.reporting.common.model.StudentFilters;
@@ -42,7 +41,7 @@ abstract class AbstractQueryConverter<Q extends AbstractAggregateReportQuery, T 
     @Override
     public Collection<AggregateReportQueryDescription<T>> convert(final Q query) {
         return query.isFilteredSubgroupQuery()
-                ? convertFilteredSubgroupsQuery((AbstractSubgroupedAggregateReportQuery) query)
+                ? convertFilteredSubgroupsQuery(query)
                 : convertBasicQuery(query);
 
     }
@@ -64,16 +63,15 @@ abstract class AbstractQueryConverter<Q extends AbstractAggregateReportQuery, T 
         return queries;
     }
 
-    private Collection<AggregateReportQueryDescription<T>> convertFilteredSubgroupsQuery(final AbstractSubgroupedAggregateReportQuery reportQueryToCovert) {
+    private Collection<AggregateReportQueryDescription<T>> convertFilteredSubgroupsQuery(final Q reportQueryToCovert) {
         final List<AggregateReportQueryDescription<T>> queries = newArrayList();
 
         //convert queries to a single subject query if needed
-        for (final Q query : unwrapToSupportedQueries((Q) reportQueryToCovert)) {
-            final AbstractSubgroupedAggregateReportQuery subgroupedQuery = (AbstractSubgroupedAggregateReportQuery) query;
+        for (final Q query : unwrapToSupportedQueries(reportQueryToCovert)) {
             //each pair of organization/dimension is represented by a separate OrganizationTypeQuery
-            for (final String subgroupsKey : subgroupedQuery.getSubgroups().keySet()) {
+            for (final String subgroupsKey : query.getSubgroups().keySet()) {
                 queries.addAll(buildOrganizationQueries(query,
-                        (T) toOrganizationQueryBuilderWithCommonAttributes(query, subgroupedQuery.getSubgroups().get(subgroupsKey))
+                        (T) toOrganizationQueryBuilderWithCommonAttributes(query, query.getSubgroups().get(subgroupsKey))
                                 .subgroupKey(subgroupsKey)
                                 .dimensionType(Overall)
                                 .build()));
@@ -85,9 +83,7 @@ abstract class AbstractQueryConverter<Q extends AbstractAggregateReportQuery, T 
     private List<AggregateReportQueryDescription<T>> buildOrganizationQueries(final Q query, final T commonQuery) {
         final List<AggregateReportQueryDescription<T>> queries = newArrayList();
         // for each level of organization add one or more builders to the result
-        if (query instanceof AbstractSubgroupedAggregateReportQuery) {
-            appendState((AbstractSubgroupedAggregateReportQuery) query, queries, commonQuery);
-        }
+        appendState(query, queries, commonQuery);
         appendDistricts(query, queries, commonQuery);
         appendSchools(query, queries, commonQuery);
         return queries;
@@ -139,7 +135,7 @@ abstract class AbstractQueryConverter<Q extends AbstractAggregateReportQuery, T 
      * @param queries     the list of {@link AggregateReportQueryDescription} to add a new description if needed
      * @param commonQuery the common attributes to be included into a newly built AggregateReportQueryDescription's query
      */
-    private void appendState(final AbstractSubgroupedAggregateReportQuery query,
+    private void appendState(final Q query,
                              final List<AggregateReportQueryDescription<T>> queries,
                              final T commonQuery) {
         if (query.isIncludeState()) {
@@ -158,8 +154,7 @@ abstract class AbstractQueryConverter<Q extends AbstractAggregateReportQuery, T 
                                  final List<AggregateReportQueryDescription<T>> queries,
                                  final T commonQuery) {
 
-        if (reportQuery instanceof AbstractSubgroupedAggregateReportQuery
-                && ((AbstractSubgroupedAggregateReportQuery) reportQuery).isIncludeAllDistricts()) {
+        if (reportQuery.isIncludeAllDistricts()) {
             final T query = (T) commonQuery.copy()
                     .organizationType(State)
                     .build();
@@ -195,8 +190,7 @@ abstract class AbstractQueryConverter<Q extends AbstractAggregateReportQuery, T 
         //schools could be selected individually or as part of a district selection
         final Set<Long> districtsWithAllSchools = newHashSet();
         if (isIncludeAllSchoolsOfDistricts(reportQuery)) {
-            final Set<Long> districtIds = reportQuery instanceof AbstractSubgroupedAggregateReportQuery
-                    && ((AbstractSubgroupedAggregateReportQuery) reportQuery).isIncludeAllDistricts()
+            final Set<Long> districtIds = reportQuery.isIncludeAllDistricts()
                     ? repositoryHelper.findAllDistrictIds()
                     : reportQuery.getDistrictIds();
             districtsWithAllSchools.addAll(districtIds);
@@ -232,20 +226,12 @@ abstract class AbstractQueryConverter<Q extends AbstractAggregateReportQuery, T 
         }
     }
 
-    private boolean isIncludeAllSchoolsOfDistricts(final AbstractAggregateReportQuery query) {
-        if (!(query instanceof AbstractSubgroupedAggregateReportQuery)) {
-            return false;
-        }
-        final AbstractSubgroupedAggregateReportQuery subgroupedQuery = (AbstractSubgroupedAggregateReportQuery) query;
-        return subgroupedQuery.isIncludeAllSchoolsOfDistricts()
-                && (!query.getDistrictIds().isEmpty() || subgroupedQuery.isIncludeAllDistricts());
+    private boolean isIncludeAllSchoolsOfDistricts(final Q query) {
+        return query.isIncludeAllSchoolsOfDistricts()
+                && (!query.getDistrictIds().isEmpty() || query.isIncludeAllDistricts());
     }
 
-    private boolean isIncludeAllDistrictsOfSchools(final AbstractAggregateReportQuery query) {
-        if (!(query instanceof AbstractSubgroupedAggregateReportQuery)) {
-            return false;
-        }
-        final AbstractSubgroupedAggregateReportQuery subgroupedQuery = (AbstractSubgroupedAggregateReportQuery) query;
-        return subgroupedQuery.isIncludeAllDistrictsOfSchools() && !query.getSchoolIds().isEmpty();
+    private boolean isIncludeAllDistrictsOfSchools(final Q query) {
+        return query.isIncludeAllDistrictsOfSchools() && !query.getSchoolIds().isEmpty();
     }
 }
