+++ aggregate-service/src/main/java/org/opentestsystem/rdw/olap/service/impl/TargetReportQueryHandler.java
@@ -93,7 +93,9 @@ class TargetReportQueryHandler extends AbstractAggregateQueryHandler<TargetRow,
         checkArgument(!query.isFilteredSubgroupQuery(), "invalid query, filtered subgroups are not supported");
         checkArgument(query.getSubjectCode() != null, "invalid query, missing a subject");
         checkArgument(query.getAssessmentGradeCodes().size() == 1, "invalid query, must have one assessment grade");
-        checkArgument(((query.getSchoolIds().size() == 1 && query.getDistrictIds().isEmpty()) ||
+        checkArgument(!query.isIncludeAllDistrictsOfSchools() && !query.isIncludeAllDistricts()
+                && !query.isIncludeState() && !query.isIncludeAllSchoolsOfDistricts()
+                && ((query.getSchoolIds().size() == 1 && query.getDistrictIds().isEmpty()) ||
                 (query.getDistrictIds().size() == 1 && query.getSchoolIds().isEmpty())), "invalid query, supports one school or one district only");
         checkArgument(!query.isFilteredSubgroupQuery(), "does not support filtered subgroups");
