+++ aggregate-service/src/main/java/org/opentestsystem/rdw/olap/sqlbuilder/impl/QueryProviderRepositoryHelper.java
@@ -2,13 +2,41 @@ package org.opentestsystem.rdw.olap.sqlbuilder.impl;
 
 import com.google.common.collect.ImmutableList;
 import org.opentestsystem.rdw.olap.model.OrganizationTypeQuery;
-import org.opentestsystem.rdw.olap.repository.*;
+import org.opentestsystem.rdw.olap.repository.ActiveAssessmentRepository;
+import org.opentestsystem.rdw.olap.repository.AdministrationConditionRepository;
+import org.opentestsystem.rdw.olap.repository.AssessmentGradeRepository;
+import org.opentestsystem.rdw.olap.repository.AssessmentTypeRepository;
+import org.opentestsystem.rdw.olap.repository.BooleanRepository;
+import org.opentestsystem.rdw.olap.repository.CodedEntityRepository;
+import org.opentestsystem.rdw.olap.repository.CompletenessRepository;
+import org.opentestsystem.rdw.olap.repository.ElasRepository;
+import org.opentestsystem.rdw.olap.repository.EthnicityRepository;
+import org.opentestsystem.rdw.olap.repository.GenderRepository;
+import org.opentestsystem.rdw.olap.repository.LanguageRepository;
+import org.opentestsystem.rdw.olap.repository.MilitaryConnectedCodeRepository;
+import org.opentestsystem.rdw.olap.repository.OrganizationRepository;
+import org.opentestsystem.rdw.olap.repository.SchoolYearRepository;
+import org.opentestsystem.rdw.olap.repository.StrictBooleanRepository;
+import org.opentestsystem.rdw.olap.repository.SubjectRepository;
+import org.opentestsystem.rdw.olap.repository.TargetRepository;
 import org.opentestsystem.rdw.olap.sqlbuilder.ReportTemplateProvider;
-import org.opentestsystem.rdw.reporting.common.model.*;
+import org.opentestsystem.rdw.reporting.common.model.ActiveAssessment;
+import org.opentestsystem.rdw.reporting.common.model.CodedEntity;
+import org.opentestsystem.rdw.reporting.common.model.Dimension;
+import org.opentestsystem.rdw.reporting.common.model.DimensionType;
+import org.opentestsystem.rdw.reporting.common.model.District;
+import org.opentestsystem.rdw.reporting.common.model.Grade;
+import org.opentestsystem.rdw.reporting.common.model.School;
+import org.opentestsystem.rdw.reporting.common.model.SubjectAssessment;
+import org.opentestsystem.rdw.reporting.common.model.Target;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.*;
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.TreeMap;
 import java.util.stream.Collectors;
 
 import static com.google.common.collect.Maps.newHashMap;
