+++ aggregate-service/src/test/java/org/opentestsystem/rdw/olap/service/impl/AbstractAggregateQueryHandlerTest.java
@@ -12,7 +12,6 @@ import org.opentestsystem.rdw.olap.service.AggregateQueryConverter;
 import org.opentestsystem.rdw.olap.service.AggregateQueryHandler;
 import org.opentestsystem.rdw.olap.service.AuthorizationService;
 import org.opentestsystem.rdw.reporting.common.model.AbstractAggregateReportQuery;
-import org.opentestsystem.rdw.reporting.common.model.AbstractSubgroupedAggregateReportQuery;
 import org.opentestsystem.rdw.reporting.common.model.AggregateRow;
 import org.opentestsystem.rdw.reporting.common.model.OrganizationType;
 import org.opentestsystem.rdw.reporting.common.web.security.User;
@@ -23,7 +22,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static org.opentestsystem.rdw.common.model.AssessmentType.ICA;
 import static org.opentestsystem.rdw.common.model.AssessmentType.SUMMATIVE;
 
 @SuppressWarnings("unchecked")
