+++ aggregate-service/src/test/java/org/opentestsystem/rdw/olap/service/impl/AbstractQueryConverterTest.java
@@ -6,7 +6,6 @@ import org.opentestsystem.rdw.olap.model.AggregateReportQueryDescription;
 import org.opentestsystem.rdw.olap.model.OrganizationTypeQuery;
 import org.opentestsystem.rdw.olap.sqlbuilder.impl.QueryProviderRepositoryHelper;
 import org.opentestsystem.rdw.reporting.common.model.AbstractAggregateReportQuery;
-import org.opentestsystem.rdw.reporting.common.model.AbstractSubgroupedAggregateReportQuery;
 import org.opentestsystem.rdw.reporting.common.model.OrganizationType;
 import org.opentestsystem.rdw.reporting.common.model.StudentFilters;
 
@@ -20,7 +19,7 @@ import static org.opentestsystem.rdw.reporting.common.model.DimensionType.Gender
 import static org.opentestsystem.rdw.reporting.common.model.DimensionType.Overall;
 
 @SuppressWarnings("unchecked")
-public abstract class AbstractQueryConverterTest<Q extends AbstractSubgroupedAggregateReportQuery, B extends AbstractSubgroupedAggregateReportQuery.Builder<Q, B>> {
+public abstract class AbstractQueryConverterTest<Q extends AbstractAggregateReportQuery, B extends AbstractAggregateReportQuery.Builder<Q, B>> {
     private AbstractQueryConverter queryConverter;
     private QueryProviderRepositoryHelper repositoryHelper;
     private final String subgroupOverall = Overall.name();
