+++ aggregate-service/src/test/java/org/opentestsystem/rdw/olap/service/impl/DefaultReportOptionServiceTest.java
@@ -20,7 +20,11 @@ import org.opentestsystem.rdw.olap.service.LanguageService;
 import org.opentestsystem.rdw.olap.service.MilitaryConnectedCodeService;
 import org.opentestsystem.rdw.olap.service.SchoolYearService;
 import org.opentestsystem.rdw.olap.service.SubjectService;
-import org.opentestsystem.rdw.reporting.common.model.*;
+import org.opentestsystem.rdw.reporting.common.model.Claim;
+import org.opentestsystem.rdw.reporting.common.model.CodedEntity;
+import org.opentestsystem.rdw.reporting.common.model.DimensionType;
+import org.opentestsystem.rdw.reporting.common.model.Grade;
+import org.opentestsystem.rdw.reporting.common.model.SubjectAssessment;
 
 import java.util.Arrays;
 import java.util.List;
