+++ aggregate-service/src/test/java/org/opentestsystem/rdw/olap/service/impl/TargetQueryConverterTest.java
@@ -1,25 +1,26 @@
 package org.opentestsystem.rdw.olap.service.impl;
 
-import org.junit.Before;
 import org.junit.Test;
+import org.opentestsystem.rdw.olap.model.OrganizationTypeQuery;
 import org.opentestsystem.rdw.olap.sqlbuilder.impl.QueryProviderRepositoryHelper;
 import org.opentestsystem.rdw.reporting.common.model.DimensionType;
+import org.opentestsystem.rdw.reporting.common.model.StudentFilters;
 import org.opentestsystem.rdw.reporting.common.model.TargetReportQuery;
 import org.opentestsystem.rdw.reporting.common.model.TargetReportQuery.Builder;
 
 import static com.google.common.collect.ImmutableSet.of;
-import static org.mockito.Mockito.mock;
 import static org.opentestsystem.rdw.common.model.AssessmentType.SUMMATIVE;
 
-// TODO expand tests
-public class TargetQueryConverterTest {
+public class TargetQueryConverterTest extends AbstractQueryConverterTest<TargetReportQuery, Builder> {
     private TargetQueryConverter queryConverter;
 
-    protected TargetQueryConverter createQueryConverterUnderTest(final QueryProviderRepositoryHelper repositoryHelper) {
+    @Override
+    protected AbstractQueryConverter createQueryConverterUnderTest(final QueryProviderRepositoryHelper repositoryHelper) {
         queryConverter = new TargetQueryConverter(repositoryHelper);
         return queryConverter;
     }
 
+    @Override
     protected Builder getAggregateReportQueryBuilder() {
         return TargetReportQuery.builder()
                 .dimensionType(DimensionType.Overall)
@@ -30,6 +31,12 @@ public class TargetQueryConverterTest {
                 .assessmentTypeCode(SUMMATIVE.code());
     }
 
+    @Override
+    protected OrganizationTypeQuery.Builder toQueryBuilderWithCommonAttributes(final TargetReportQuery reportQuery,
+                                                                               final StudentFilters filters) {
+        return AggregateTestUtils.toQueryClaimBuilderWithCommonAttributes(reportQuery, filters);
+    }
+
     @Test(expected = IllegalArgumentException.class)
     public void itShouldFailForQueryWithMultipleGrades() {
         final TargetReportQuery query = getAggregateReportQueryBuilder()
@@ -38,15 +45,4 @@ public class TargetQueryConverterTest {
 
         queryConverter.getOrganizationTypeQueryBuilder(query, query.getStudentFilters());
     }
-
-    // TODO copy pasted from AbstractQueryConverterTest but adjusted to remove subgroups - note this is an intermediate fix
-
-    private QueryProviderRepositoryHelper repositoryHelper;
-
-    @Before
-    public void setUp() {
-        repositoryHelper = mock(QueryProviderRepositoryHelper.class);
-        queryConverter = createQueryConverterUnderTest(repositoryHelper);
-    }
-
 }
