+++ aggregate-service/src/test/java/org/opentestsystem/rdw/olap/service/impl/TargetReportQueryHandlerTest.java
@@ -58,7 +58,7 @@ public class TargetReportQueryHandlerTest extends AbstractAggregateQueryHandlerT
         repository = mock(TargetReportRepository.class);
         subjectService = mock(SubjectService.class);
         when(subjectService.getSubjects()).thenReturn(newArrayList(
-           SubjectAssessment.builder().id(1).code("Math").assessmentType("sum").targetReport(true).build()
+                SubjectAssessment.builder().id(1).code("Math").assessmentType("sum").targetReport(true).build()
         ));
         serviceUnderTest = new TargetReportQueryHandler(repository, authorizationService, aggregateReportSettings, queryConverter, threadPoolTaskExecutor, subjectService, 0);
         return serviceUnderTest;
@@ -160,6 +160,13 @@ public class TargetReportQueryHandlerTest extends AbstractAggregateQueryHandlerT
     }
 
     @Test
+    public void itShouldFailWithState() {
+        thrown.expect(IllegalArgumentException.class);
+        thrown.expectMessage(containsString("one school or one district only"));
+        serviceUnderTest.estimateReportRowCount(user, getAggregateQueryBuilderWithValidQuery().includeState(true).build());
+    }
+
+    @Test
     public void itShouldFailWithMoreThanOneSchool() {
         thrown.expect(IllegalArgumentException.class);
         thrown.expectMessage(containsString("one school or one district only"));
@@ -178,6 +185,34 @@ public class TargetReportQueryHandlerTest extends AbstractAggregateQueryHandlerT
         thrown.expect(IllegalArgumentException.class);
         thrown.expectMessage(containsString("one school or one district only"));
         serviceUnderTest.estimateReportRowCount(user, getAggregateQueryBuilderWithValidQuery().districtIds(of(123L)).build());
+    }
+
+    @Test
+    public void itShouldFailWithSchoolAndItsDistricts() {
+        thrown.expect(IllegalArgumentException.class);
+        thrown.expectMessage(containsString("one school or one district only"));
+        serviceUnderTest.estimateReportRowCount(user, getAggregateQueryBuilderWithValidQuery().includeAllSchoolsOfDistricts(true).build());
+    }
+
+    @Test
+    public void itShouldFailWithDistrictAndItsSchools() {
+        thrown.expect(IllegalArgumentException.class);
+        thrown.expectMessage(containsString("one school or one district only"));
+        serviceUnderTest.estimateReportRowCount(user, getAggregateQueryBuilderWithValidQuery().districtIds(of(124L)).schoolIds(of()).includeAllSchoolsOfDistricts(true).build());
+    }
+
+    @Test
+    public void itShouldFailWithAllDistricts() {
+        thrown.expect(IllegalArgumentException.class);
+        thrown.expectMessage(containsString("one school or one district only"));
+        serviceUnderTest.estimateReportRowCount(user, getAggregateQueryBuilderWithValidQuery().includeAllDistricts(true).build());
+    }
+
+    @Test
+    public void itShouldFailWithFilteredSubgroups() {
+        thrown.expect(IllegalArgumentException.class);
+        thrown.expectMessage(containsString("filtered subgroups are not supported"));
+        serviceUnderTest.estimateReportRowCount(user, getAggregateQueryBuilderWithValidQuery().subgroup("test", StudentFilters.builder().build()).build());
     }
 
     @Test
