+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/model/AbstractAggregateReportQuery.java
@@ -5,8 +5,10 @@ import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
 
+import static com.google.common.collect.Maps.newHashMap;
 import static com.google.common.collect.Sets.newHashSet;
 import static org.opentestsystem.rdw.reporting.common.util.CollectionUtils.safeImmutableOf;
 
@@ -15,61 +17,71 @@ import static org.opentestsystem.rdw.reporting.common.util.CollectionUtils.safeI
  */
 @SuppressWarnings("unchecked")
 public abstract class AbstractAggregateReportQuery extends AbstractReportQuery implements AggregateQuery {
+
+    protected Set<String> subjectCodes;
     protected String assessmentTypeCode;
     protected ImmutableSet<String> assessmentGradeCodes;
+    protected boolean includeState;
+    protected boolean includeAllDistricts;
+    protected boolean includeAllSchoolsOfDistricts;
+    protected boolean includeAllDistrictsOfSchools;
     protected ImmutableSet<Long> districtIds;
     protected ImmutableSet<Long> schoolIds;
     protected ImmutableSet<String> completenessCodes;
     protected ImmutableSet<String> administrativeConditionCodes;
 
-    //pass through values
+    // display preferences used by the UI to present the report
     protected boolean showEmpty = true;
     protected String valueDisplayType;
     protected String achievementLevelDisplayType;
     protected ImmutableList<String> columnOrder;
 
-    //defines custom subgroups
-    protected ImmutableMap<String, StudentFilters> subgroups;
-
-    //not supported with the subgroups
+    // use of these is mutually exclusive with subgroups
     protected ImmutableSet<DimensionType> dimensionTypes;
     protected StudentFilters studentFilters;
 
-    public abstract boolean isFilteredSubgroupQuery();
-
-    public Set<DimensionType> getDimensionTypes() {
-        return safeImmutableOf(dimensionTypes);
-    }
-
-    public StudentFilters getStudentFilters() {
-        return studentFilters == null ? new StudentFilters() : studentFilters;
-    }
+    // defines custom subgroups
+    protected Map<String, StudentFilters> subgroups;
 
     @Override
+    public boolean isFilteredSubgroupQuery() {
+        return !getSubgroups().isEmpty();
+    }
+
+    public Set<String> getSubjectCodes() {
+        return safeImmutableOf(subjectCodes);
+    }
+
     public String getAssessmentTypeCode() {
         return assessmentTypeCode;
-    }
-
-    @Override
-    public Set<Long> getDistrictIds() {
-        return safeImmutableOf(districtIds);
-    }
-
-    @Override
-    public Set<Long> getSchoolIds() {
-        return safeImmutableOf(schoolIds);
     }
 
     public Set<String> getAssessmentGradeCodes() {
         return safeImmutableOf(assessmentGradeCodes);
     }
 
-    public boolean isShowEmpty() {
-        return showEmpty;
+    public boolean isIncludeState() {
+        return includeState;
     }
 
-    public List<String> getColumnOrder() {
-        return safeImmutableOf(columnOrder);
+    public boolean isIncludeAllDistricts() {
+        return includeAllDistricts;
+    }
+
+    public boolean isIncludeAllSchoolsOfDistricts() {
+        return includeAllSchoolsOfDistricts;
+    }
+
+    public boolean isIncludeAllDistrictsOfSchools() {
+        return includeAllDistrictsOfSchools;
+    }
+
+    public Set<Long> getDistrictIds() {
+        return safeImmutableOf(districtIds);
+    }
+
+    public Set<Long> getSchoolIds() {
+        return safeImmutableOf(schoolIds);
     }
 
     public Set<String> getCompletenessCodes() {
@@ -80,46 +92,104 @@ public abstract class AbstractAggregateReportQuery extends AbstractReportQuery i
         return safeImmutableOf(administrativeConditionCodes);
     }
 
+    public boolean isShowEmpty() {
+        return showEmpty;
+    }
+
+    public String getValueDisplayType() {
+        return valueDisplayType;
+    }
+
+    public String getAchievementLevelDisplayType() {
+        return achievementLevelDisplayType;
+    }
+
+    public List<String> getColumnOrder() {
+        return safeImmutableOf(columnOrder);
+    }
+
+    public Set<DimensionType> getDimensionTypes() {
+        return safeImmutableOf(dimensionTypes);
+    }
+
+    public StudentFilters getStudentFilters() {
+        return studentFilters == null ? new StudentFilters() : studentFilters;
+    }
+
+    public Map<String, StudentFilters> getSubgroups() {
+        return safeImmutableOf(subgroups);
+    }
+
     public static abstract class Builder<A extends AbstractAggregateReportQuery, B extends Builder<A, B>> extends AbstractReportQuery.Builder<A, B> {
 
+        private Set<String> subjectCodes;
         private String assessmentTypeCode;
         private Set<String> assessmentGradeCodes;
+        private boolean includeState;
+        private boolean includeAllDistricts;
+        private boolean includeAllSchoolsOfDistricts;
+        private boolean includeAllDistrictsOfSchools;
         private Set<Long> districtIds;
         private Set<Long> schoolIds;
         private Set<String> completenessCodes;
         private Set<String> administrativeConditionCodes;
         private boolean showEmpty = true;
+        private String valueDisplayType;
+        private String achievementLevelDisplayType;
         private List<String> columnOrder;
         private Set<DimensionType> dimensionTypes;
         private StudentFilters studentFilters;
+        private Map<String, StudentFilters> subgroups;
 
         protected A build(A value) {
             super.build(value);
+            value.subjectCodes = subjectCodes;
             value.assessmentTypeCode = assessmentTypeCode;
             value.assessmentGradeCodes = safeImmutableOf(assessmentGradeCodes);
+            value.includeState = includeState;
+            value.includeAllDistricts = includeAllDistricts;
+            value.includeAllSchoolsOfDistricts = includeAllSchoolsOfDistricts;
+            value.includeAllDistrictsOfSchools = includeAllDistrictsOfSchools;
             value.schoolIds = safeImmutableOf(schoolIds);
             value.districtIds = safeImmutableOf(districtIds);
             value.completenessCodes = safeImmutableOf(completenessCodes);
             value.administrativeConditionCodes = safeImmutableOf(administrativeConditionCodes);
             value.showEmpty = showEmpty;
+            value.valueDisplayType = valueDisplayType;
+            value.achievementLevelDisplayType = achievementLevelDisplayType;
             value.columnOrder = safeImmutableOf(columnOrder);
             value.dimensionTypes = dimensionTypes == null ? null : ImmutableSet.copyOf(dimensionTypes);
             value.studentFilters = studentFilters;
+            value.subgroups = subgroups == null ? null : ImmutableMap.copyOf(subgroups);
             return value;
         }
 
         protected B copy(final A value) {
             return super.copy(value)
+                    .subjectCodes(value.subjectCodes)
                     .assessmentTypeCode(value.assessmentTypeCode)
                     .assessmentGradeCodes(value.assessmentGradeCodes)
+                    .includeState(value.includeState)
+                    .includeAllDistricts(value.includeAllDistricts)
+                    .includeAllSchoolsOfDistricts(value.includeAllSchoolsOfDistricts)
+                    .includeAllDistrictsOfSchools(value.includeAllDistrictsOfSchools)
                     .schoolIds(value.schoolIds)
                     .districtIds(value.districtIds)
                     .completenessCodes(value.completenessCodes)
                     .administrativeConditionCodes(value.administrativeConditionCodes)
                     .showEmpty(value.showEmpty)
+                    .valueDisplayType(value.valueDisplayType)
+                    .achievementLevelDisplayType(value.achievementLevelDisplayType)
                     .columnOrder(value.columnOrder)
                     .dimensionTypes(value.dimensionTypes == null ? null : newHashSet(value.dimensionTypes)) // TODO redundant?
-                    .studentFilters(value.studentFilters);
+                    .studentFilters(value.studentFilters)
+                    .subgroups(newHashMap(value.getSubgroups()));
+        }
+
+
+        public B subjectCodes(final Set<String> subjectCodes) {
+            this.subjectCodes = subjectCodes;
+            return (B) this;
         }
 
         public B assessmentTypeCode(final String assessmentTypeCode) {
@@ -129,6 +199,26 @@ public abstract class AbstractAggregateReportQuery extends AbstractReportQuery i
 
         public B assessmentGradeCodes(final Set<String> assessmentGradeCodes) {
             this.assessmentGradeCodes = assessmentGradeCodes;
+            return (B) this;
+        }
+
+        public B includeState(final boolean includeState) {
+            this.includeState = includeState;
+            return (B) this;
+        }
+
+        public B includeAllDistricts(final boolean includeAllDistricts) {
+            this.includeAllDistricts = includeAllDistricts;
+            return (B) this;
+        }
+
+        public B includeAllSchoolsOfDistricts(final boolean includeAllSchoolsOfDistricts) {
+            this.includeAllSchoolsOfDistricts = includeAllSchoolsOfDistricts;
+            return (B) this;
+        }
+
+        public B includeAllDistrictsOfSchools(final boolean includeAllDistrictsOfSchools) {
+            this.includeAllDistrictsOfSchools = includeAllDistrictsOfSchools;
             return (B) this;
         }
 
@@ -157,6 +247,16 @@ public abstract class AbstractAggregateReportQuery extends AbstractReportQuery i
             return (B) this;
         }
 
+        public B valueDisplayType(final String valueDisplayType) {
+            this.valueDisplayType = valueDisplayType;
+            return (B) this;
+        }
+
+        public B achievementLevelDisplayType(final String achievementLevelDisplayType) {
+            this.achievementLevelDisplayType = achievementLevelDisplayType;
+            return (B) this;
+        }
+
         public B columnOrder(final List<String> columnOrder) {
             this.columnOrder = columnOrder;
             return (B) this;
@@ -175,6 +275,19 @@ public abstract class AbstractAggregateReportQuery extends AbstractReportQuery i
 
         public B studentFilters(final StudentFilters studentFilters) {
             this.studentFilters = studentFilters;
+            return (B) this;
+        }
+
+        public B subgroups(final Map<String, StudentFilters> subgroups) {
+            this.subgroups = subgroups;
+            return (B) this;
+        }
+
+        public B subgroup(final String label, final StudentFilters filters) {
+            if (this.subgroups == null) {
+                this.subgroups = newHashMap();
+            }
+            this.subgroups.put(label, filters);
             return (B) this;
         }
