+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/model/AbstractReportQuery.java
@@ -8,8 +8,6 @@ public abstract class AbstractReportQuery implements ReportQuery {
         return name;
     }
 
-    public abstract ReportQueryType getType();
-
     public abstract <T> T copy();
 
     @SuppressWarnings({"unchecked"})
