+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/model/ClaimAggregateReportQuery.java
@@ -14,7 +14,7 @@ import static org.opentestsystem.rdw.reporting.common.util.CollectionUtils.safeI
  */
 @SuppressWarnings("unchecked")
 @JsonIgnoreProperties(value="subjectCodes", allowGetters = true)
-public class ClaimAggregateReportQuery extends AbstractSubgroupedAggregateReportQuery {
+public class ClaimAggregateReportQuery extends AbstractAggregateReportQuery {
     private ImmutableSet<Integer> schoolYears;
     private ImmutableMap<String, Set<String>> claimCodesBySubject;
 
@@ -22,7 +22,7 @@ public class ClaimAggregateReportQuery extends AbstractSubgroupedAggregateReport
         return ReportQueryType.Claim;
     }
 
-    // TODO remove?
+    // TODO move schoolYears to base report & enforce validation where single year is required
     @Override
     public Set<Integer> getSchoolYears() {
         return safeImmutableOf(schoolYears);
@@ -32,20 +32,21 @@ public class ClaimAggregateReportQuery extends AbstractSubgroupedAggregateReport
         return claimCodesBySubject == null ? ImmutableMap.of() : claimCodesBySubject;
     }
 
+    // TODO do we need this?
     @Override
     public Set<String> getSubjectCodes() {
         return getClaimCodesBySubject().keySet();
     }
 
     public Builder copy() {
-        return new Builder().copy(this);
+        return builder().copy(this);
     }
 
     public static Builder builder() {
         return new Builder();
     }
 
-    public static class Builder extends AbstractSubgroupedAggregateReportQuery.Builder<ClaimAggregateReportQuery, Builder> {
+    public static class Builder extends AbstractAggregateReportQuery.Builder<ClaimAggregateReportQuery, Builder> {
         private Set<Integer> schoolYears;
         private Map<String, Set<String>> claimCodesBySubject;
