+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/model/CustomAggregateReportQuery.java
@@ -9,7 +9,7 @@ import static org.opentestsystem.rdw.reporting.common.util.CollectionUtils.safeI
 /**
  * This class represents a query for the aggregate report with {@link AggregateRow}s
  */
-public class CustomAggregateReportQuery extends AbstractSubgroupedAggregateReportQuery {
+public class CustomAggregateReportQuery extends AbstractAggregateReportQuery {
     private ImmutableSet<Integer> schoolYears;
 
     @Override
@@ -17,20 +17,20 @@ public class CustomAggregateReportQuery extends AbstractSubgroupedAggregateRepor
         return ReportQueryType.CustomAggregate;
     }
 
-    // TODO remove
+    // TODO move schoolYears to base report & enforce validation where single year is required
     public Set<Integer> getSchoolYears() {
         return safeImmutableOf(schoolYears);
     }
 
     public Builder copy() {
-        return new Builder().copy(this);
+        return builder().copy(this);
     }
 
     public static Builder builder() {
         return new Builder();
     }
 
-    public static class Builder extends AbstractSubgroupedAggregateReportQuery.Builder<CustomAggregateReportQuery, Builder> {
+    public static class Builder extends AbstractAggregateReportQuery.Builder<CustomAggregateReportQuery, Builder> {
         private Set<Integer> schoolYears;
 
         @Override
