+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/model/DistrictSchoolExportReportQuery.java
@@ -33,7 +33,7 @@ public class DistrictSchoolExportReportQuery extends AbstractExamReportQuery {
     }
 
     public Builder copy() {
-        return new Builder().copy(this);
+        return builder().copy(this);
     }
 
     public static Builder builder() {
