+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/model/GroupPrintableReportQuery.java
@@ -14,7 +14,7 @@ public class GroupPrintableReportQuery extends AbstractBatchPrintableReportQuery
     }
 
     public Builder copy() {
-        return new Builder().copy(this);
+        return builder().copy(this);
     }
 
     public static Builder builder() {
