+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/model/LongitudinalReportQuery.java
@@ -5,13 +5,11 @@ import com.google.common.collect.ImmutableSet;
 
 import java.util.Set;
 
-import static org.opentestsystem.rdw.reporting.common.util.CollectionUtils.safeImmutableOf;
-
 /**
  * This class represents a query for the aggregate report with {@link LongitudinalRow}s
  */
 @SuppressWarnings("unchecked")
-public class LongitudinalReportQuery extends AbstractSubgroupedAggregateReportQuery {
+public class LongitudinalReportQuery extends AbstractAggregateReportQuery {
     private int toSchoolYear;
 
     @Override
@@ -31,14 +29,14 @@ public class LongitudinalReportQuery extends AbstractSubgroupedAggregateReportQu
     }
 
     public Builder copy() {
-        return new Builder().copy(this);
+        return builder().copy(this);
     }
 
     public static Builder builder() {
         return new Builder();
     }
 
-    public static class Builder extends AbstractSubgroupedAggregateReportQuery.Builder<LongitudinalReportQuery, Builder> {
+    public static class Builder extends AbstractAggregateReportQuery.Builder<LongitudinalReportQuery, Builder> {
         private int toSchoolYear;
 
         @Override
