+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/model/SchoolGradePrintableReportQuery.java
@@ -20,7 +20,7 @@ public class SchoolGradePrintableReportQuery extends AbstractBatchPrintableRepor
     }
 
     public Builder copy() {
-        return new Builder().copy(this);
+        return builder().copy(this);
     }
 
     public static Builder builder() {
