+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/model/StudentPrintableReportQuery.java
@@ -14,7 +14,7 @@ public class StudentPrintableReportQuery extends AbstractPrintableReportQuery {
     }
 
     public Builder copy() {
-        return new Builder().copy(this);
+        return builder().copy(this);
     }
 
     public static Builder builder() {
