+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/model/SubjectDefinition.java
@@ -45,7 +45,7 @@ public class SubjectDefinition {
     }
 
     public Builder copy() {
-        return new Builder().copy(this);
+        return builder().copy(this);
     }
 
     public static Builder builder() {
