+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/model/TargetReportQuery.java
@@ -45,7 +45,7 @@ public class TargetReportQuery extends AbstractAggregateReportQuery {
     }
 
     public Builder copy() {
-        return new Builder().copy(this);
+        return builder().copy(this);
     }
 
     public static Builder builder() {
