+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/model/TargetRow.java
@@ -33,7 +33,7 @@ public class TargetRow extends AggregateRow {
     }
 
     public Builder copy() {
-        return new Builder().copy(this);
+        return builder().copy(this);
     }
 
     @SuppressWarnings("unchecked")
