+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/model/jackson/ReportQueryMixin.java
@@ -1,9 +1,15 @@
 package org.opentestsystem.rdw.reporting.common.model.jackson;
 
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
-import org.opentestsystem.rdw.reporting.common.model.*;
+import org.opentestsystem.rdw.reporting.common.model.ClaimAggregateReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.CustomAggregateReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.DistrictSchoolExportReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.GroupPrintableReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.LongitudinalReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.SchoolGradePrintableReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.StudentPrintableReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.TargetReportQuery;
 
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.PROPERTY, property = "type")
 @JsonSubTypes({
