+++ report-processor/src/main/java/org/opentestsystem/rdw/reporting/processor/model/UserReport.java
@@ -86,7 +86,7 @@ public class UserReport {
      * @return A Builder initialized with this UserReport's values.
      */
     public Builder copy() {
-        return new Builder().copy(this);
+        return builder().copy(this);
     }
 
     /**
