+++ report-processor/src/main/java/org/opentestsystem/rdw/reporting/processor/service/DefaultExamReportService.java
@@ -7,10 +7,8 @@ import org.opentestsystem.rdw.reporting.common.model.ReportQueryType;
 import org.opentestsystem.rdw.reporting.common.web.security.User;
 import org.opentestsystem.rdw.reporting.processor.model.UserReport;
 import org.opentestsystem.rdw.reporting.processor.web.ReportContentResource;
-import org.opentestsystem.rdw.security.service.PermissionServiceException;
-import org.springframework.security.access.AuthorizationServiceException;
-import org.springframework.stereotype.Service;
 import org.opentestsystem.rdw.security.Permission;
+import org.springframework.stereotype.Service;
 
 import javax.validation.constraints.NotNull;
 import java.util.List;
@@ -18,13 +16,16 @@ import java.util.Map;
 import java.util.NoSuchElementException;
 import java.util.Set;
 
-import static org.opentestsystem.rdw.reporting.common.model.ReportQueryType.*;
+import static org.opentestsystem.rdw.reporting.common.model.ReportQueryType.Claim;
+import static org.opentestsystem.rdw.reporting.common.model.ReportQueryType.CustomAggregate;
+import static org.opentestsystem.rdw.reporting.common.model.ReportQueryType.DistrictSchoolExport;
+import static org.opentestsystem.rdw.reporting.common.model.ReportQueryType.Group;
+import static org.opentestsystem.rdw.reporting.common.model.ReportQueryType.Longitudinal;
+import static org.opentestsystem.rdw.reporting.common.model.ReportQueryType.SchoolGrade;
+import static org.opentestsystem.rdw.reporting.common.model.ReportQueryType.Student;
+import static org.opentestsystem.rdw.reporting.common.model.ReportQueryType.Target;
 import static org.opentestsystem.rdw.reporting.common.security.ReportingPermission.GroupPiiRead;
 import static org.opentestsystem.rdw.reporting.common.security.ReportingPermission.IndividualPiiRead;
-
-
-
-
 
 
 @Service
