+++ report-processor/src/main/java/org/opentestsystem/rdw/reporting/processor/stream/FetchStudents.java
@@ -2,11 +2,17 @@ package org.opentestsystem.rdw.reporting.processor.stream;
 
 import com.google.common.collect.Lists;
 import com.google.common.collect.Ordering;
-import org.opentestsystem.rdw.reporting.common.model.*;
+import org.opentestsystem.rdw.reporting.common.model.AbstractBatchPrintableReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.PrintableReportOrder;
+import org.opentestsystem.rdw.reporting.common.model.ReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.ReportStatus;
+import org.opentestsystem.rdw.reporting.common.model.Student;
 import org.opentestsystem.rdw.reporting.common.stream.MessageSecurityService;
 import org.opentestsystem.rdw.reporting.common.web.security.User;
 import org.opentestsystem.rdw.reporting.processor.configuration.ReportGenerationProperties;
-import org.opentestsystem.rdw.reporting.processor.model.*;
+import org.opentestsystem.rdw.reporting.processor.model.ExamQueryParams;
+import org.opentestsystem.rdw.reporting.processor.model.ProcessStudentChunkMessage;
+import org.opentestsystem.rdw.reporting.processor.model.ReportRequestMessage;
 import org.opentestsystem.rdw.reporting.processor.model.UserReport;
 import org.opentestsystem.rdw.reporting.processor.repository.StudentRepository;
 import org.opentestsystem.rdw.reporting.processor.requesthandler.ExamQueryParamsParser;
