+++ report-processor/src/main/java/org/opentestsystem/rdw/reporting/processor/stream/StudentReportProcessor.java
@@ -3,11 +3,21 @@ package org.opentestsystem.rdw.reporting.processor.stream;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Ordering;
 import org.opentestsystem.rdw.common.model.AssessmentType;
-import org.opentestsystem.rdw.reporting.common.model.*;
+import org.opentestsystem.rdw.reporting.common.model.AbstractPrintableReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.ReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.ReportStatus;
+import org.opentestsystem.rdw.reporting.common.model.Student;
+import org.opentestsystem.rdw.reporting.common.model.SubjectDefinition;
 import org.opentestsystem.rdw.reporting.common.service.SubjectDefinitionService;
 import org.opentestsystem.rdw.reporting.common.stream.MessageSecurityService;
 import org.opentestsystem.rdw.reporting.common.web.security.User;
-import org.opentestsystem.rdw.reporting.processor.model.*;
+import org.opentestsystem.rdw.reporting.processor.model.AbstractExamReport;
+import org.opentestsystem.rdw.reporting.processor.model.ExamQueryParams;
+import org.opentestsystem.rdw.reporting.processor.model.IabReport;
+import org.opentestsystem.rdw.reporting.processor.model.IcaSummativeReport;
+import org.opentestsystem.rdw.reporting.processor.model.ProcessStudentChunkMessage;
+import org.opentestsystem.rdw.reporting.processor.model.ReportRequestMessage;
+import org.opentestsystem.rdw.reporting.processor.model.StudentCompositeReport;
 import org.opentestsystem.rdw.reporting.processor.model.UserReport;
 import org.opentestsystem.rdw.reporting.processor.repository.IabReportRepository;
 import org.opentestsystem.rdw.reporting.processor.repository.IcaSummativeReportRepository;
