+++ report-processor/src/main/java/org/opentestsystem/rdw/reporting/processor/web/converter/AbstractAggregateReportCsvHandler.java
@@ -10,7 +10,14 @@ import com.google.common.base.Joiner;
 import org.apache.commons.csv.CSVFormat;
 import org.apache.commons.csv.CSVPrinter;
 import org.opentestsystem.rdw.common.model.AssessmentBoundaries;
-import org.opentestsystem.rdw.reporting.common.model.*;
+import org.opentestsystem.rdw.reporting.common.model.AbstractAggregateReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.AggregateRow;
+import org.opentestsystem.rdw.reporting.common.model.Dimension;
+import org.opentestsystem.rdw.reporting.common.model.DimensionType;
+import org.opentestsystem.rdw.reporting.common.model.Measures;
+import org.opentestsystem.rdw.reporting.common.model.ReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.ReportQueryType;
+import org.opentestsystem.rdw.reporting.common.model.StudentFilters;
 import org.opentestsystem.rdw.reporting.common.util.Headers;
 import org.opentestsystem.rdw.reporting.processor.web.ReportContentResource;
 import org.slf4j.Logger;
@@ -233,7 +240,7 @@ abstract class AbstractAggregateReportCsvHandler<Q extends AbstractAggregateRepo
 
     private String getSubgroupAsString(final CsvContext context, final String subgroupKey) {
 
-        final StudentFilters subGroupStudentFilters = ((AbstractSubgroupedAggregateReportQuery) context.getQuery())
+        final StudentFilters subGroupStudentFilters = context.getQuery()
                 .getSubgroups().get(subgroupKey);
         if (subGroupStudentFilters != null) return AttributeJoiner.join(getFiltersAsAttributes(subGroupStudentFilters));
