+++ report-processor/src/main/java/org/opentestsystem/rdw/reporting/processor/web/converter/ClaimAggregateReportCsvHandler.java
@@ -2,7 +2,11 @@ package org.opentestsystem.rdw.reporting.processor.web.converter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.opentestsystem.rdw.common.model.AssessmentBoundaries;
-import org.opentestsystem.rdw.reporting.common.model.*;
+import org.opentestsystem.rdw.reporting.common.model.ClaimAggregateReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.ClaimRow;
+import org.opentestsystem.rdw.reporting.common.model.Measures;
+import org.opentestsystem.rdw.reporting.common.model.ReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.ReportQueryType;
 import org.opentestsystem.rdw.reporting.processor.web.ReportContentResource;
 import org.springframework.context.MessageSource;
 import org.springframework.http.MediaType;
