+++ report-processor/src/main/java/org/opentestsystem/rdw/reporting/processor/web/converter/CustomAggregateReportCsvHandler.java
@@ -1,7 +1,10 @@
 package org.opentestsystem.rdw.reporting.processor.web.converter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.opentestsystem.rdw.reporting.common.model.*;
+import org.opentestsystem.rdw.reporting.common.model.AggregateRow;
+import org.opentestsystem.rdw.reporting.common.model.CustomAggregateReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.ReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.ReportQueryType;
 import org.opentestsystem.rdw.reporting.processor.web.ReportContentResource;
 import org.springframework.context.MessageSource;
 import org.springframework.http.MediaType;
