+++ report-processor/src/main/java/org/opentestsystem/rdw/reporting/processor/web/converter/TargetAggregateReportCsvHandler.java
@@ -1,7 +1,11 @@
 package org.opentestsystem.rdw.reporting.processor.web.converter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.opentestsystem.rdw.reporting.common.model.*;
+import org.opentestsystem.rdw.reporting.common.model.Measures;
+import org.opentestsystem.rdw.reporting.common.model.ReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.ReportQueryType;
+import org.opentestsystem.rdw.reporting.common.model.TargetReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.TargetRow;
 import org.opentestsystem.rdw.reporting.processor.web.ReportContentResource;
 import org.springframework.context.MessageSource;
 import org.springframework.http.MediaType;
