+++ report-processor/src/test/java/org/opentestsystem/rdw/reporting/processor/service/SpringStreamReportGeneratorTest.java
@@ -24,7 +24,9 @@ import java.time.Instant;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 import static org.opentestsystem.rdw.reporting.common.model.ReportStatus.FAILED;
 import static org.opentestsystem.rdw.reporting.common.model.ReportStatus.PENDING;
