+++ report-processor/src/test/java/org/opentestsystem/rdw/reporting/processor/stream/FetchStudentsTest.java
@@ -14,7 +14,9 @@ import org.opentestsystem.rdw.reporting.common.model.Student;
 import org.opentestsystem.rdw.reporting.common.stream.MessageSecurityService;
 import org.opentestsystem.rdw.reporting.common.web.security.User;
 import org.opentestsystem.rdw.reporting.processor.configuration.ReportGenerationProperties;
-import org.opentestsystem.rdw.reporting.processor.model.*;
+import org.opentestsystem.rdw.reporting.processor.model.ExamQueryParams;
+import org.opentestsystem.rdw.reporting.processor.model.ProcessStudentChunkMessage;
+import org.opentestsystem.rdw.reporting.processor.model.ReportRequestMessage;
 import org.opentestsystem.rdw.reporting.processor.model.UserReport;
 import org.opentestsystem.rdw.reporting.processor.repository.StudentRepository;
 import org.opentestsystem.rdw.reporting.processor.requesthandler.ExamQueryParamsParser;
@@ -42,7 +44,6 @@ import static org.mockito.Mockito.when;
 import static org.mockito.internal.verification.VerificationModeFactory.times;
 import static org.opentestsystem.rdw.reporting.common.model.ReportStatus.NO_RESULTS;
 import static org.opentestsystem.rdw.reporting.processor.MockBuilder.student;
-import static org.opentestsystem.rdw.reporting.processor.model.ExamReportOrder.STUDENT_SSID;
 import static org.opentestsystem.rdw.reporting.processor.stream.FetchStudents.TotalChunk;
 
 @RunWith(MockitoJUnitRunner.class)
