+++ report-processor/src/test/java/org/opentestsystem/rdw/reporting/processor/stream/StudentReportProcessorTest.java
@@ -8,13 +8,17 @@ import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
-import org.opentestsystem.rdw.reporting.common.model.AbstractBatchPrintableReportQuery;
 import org.opentestsystem.rdw.reporting.common.model.AbstractPrintableReportQuery;
 import org.opentestsystem.rdw.reporting.common.model.SubjectDefinition;
 import org.opentestsystem.rdw.reporting.common.service.SubjectDefinitionService;
 import org.opentestsystem.rdw.reporting.common.stream.MessageSecurityService;
 import org.opentestsystem.rdw.reporting.common.web.security.User;
-import org.opentestsystem.rdw.reporting.processor.model.*;
+import org.opentestsystem.rdw.reporting.processor.model.AbstractExamReport;
+import org.opentestsystem.rdw.reporting.processor.model.ExamQueryParams;
+import org.opentestsystem.rdw.reporting.processor.model.IabReport;
+import org.opentestsystem.rdw.reporting.processor.model.IcaSummativeReport;
+import org.opentestsystem.rdw.reporting.processor.model.ProcessStudentChunkMessage;
+import org.opentestsystem.rdw.reporting.processor.model.StudentCompositeReport;
 import org.opentestsystem.rdw.reporting.processor.model.UserReport;
 import org.opentestsystem.rdw.reporting.processor.repository.IabReportRepository;
 import org.opentestsystem.rdw.reporting.processor.repository.IcaSummativeReportRepository;
@@ -385,7 +389,7 @@ public class StudentReportProcessorTest {
         return ImmutableList.of(mathReport, elaReport);
     }
 
-    private Collection<SubjectDefinition> subjectDefinitions(final String ...subjectCodes) {
+    private Collection<SubjectDefinition> subjectDefinitions(final String... subjectCodes) {
         return Arrays.stream(subjectCodes)
                 .flatMap(subject -> AssessmentTypes.stream()
                         .map(asmtType -> subjectDefinition(subject, asmtType)))
