+++ report-processor/src/test/java/org/opentestsystem/rdw/reporting/processor/web/ReportControllerIT.java
@@ -6,13 +6,19 @@ import com.google.common.collect.ImmutableSet;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
-import org.opentestsystem.rdw.reporting.common.model.*;
+import org.opentestsystem.rdw.reporting.common.model.CustomAggregateReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.GroupPrintableReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.PrintableReportOrder;
+import org.opentestsystem.rdw.reporting.common.model.ReportStatus;
+import org.opentestsystem.rdw.reporting.common.model.SchoolGradePrintableReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.StudentPrintableReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.UserGroupId;
+import org.opentestsystem.rdw.reporting.common.model.jackson.ReportModule;
 import org.opentestsystem.rdw.reporting.common.test.support.ReportQueries;
 import org.opentestsystem.rdw.reporting.common.web.DefaultJsonContentConfiguration;
 import org.opentestsystem.rdw.reporting.common.web.WebExceptionHandlerConfiguration;
 import org.opentestsystem.rdw.reporting.common.web.security.User;
 import org.opentestsystem.rdw.reporting.processor.model.UserReport;
-import org.opentestsystem.rdw.reporting.common.model.jackson.ReportModule;
 import org.opentestsystem.rdw.reporting.processor.service.ExamReportService;
 import org.opentestsystem.rdw.reporting.processor.web.converter.ReportHandler;
 import org.springframework.beans.factory.annotation.Autowired;
