+++ report-processor/src/test/java/org/opentestsystem/rdw/reporting/processor/web/converter/AbstractAggregateReportCsvHandlerWithBasicQueryTest.java
@@ -9,7 +9,14 @@ import org.apache.commons.csv.CSVRecord;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
-import org.opentestsystem.rdw.reporting.common.model.*;
+import org.opentestsystem.rdw.reporting.common.model.AbstractAggregateReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.AggregateRow;
+import org.opentestsystem.rdw.reporting.common.model.District;
+import org.opentestsystem.rdw.reporting.common.model.Organization;
+import org.opentestsystem.rdw.reporting.common.model.ReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.ReportStatus;
+import org.opentestsystem.rdw.reporting.common.model.SchoolGradePrintableReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.StudentFilters;
 import org.opentestsystem.rdw.reporting.processor.model.UserReport;
 import org.opentestsystem.rdw.reporting.processor.web.ReportContentResource;
 import org.springframework.context.MessageSource;
