+++ report-processor/src/test/java/org/opentestsystem/rdw/reporting/processor/web/converter/AbstractAggregateReportCsvHandlerWithFilteredSubgroupQueryTest.java
@@ -9,7 +9,13 @@ import org.apache.commons.csv.CSVRecord;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
-import org.opentestsystem.rdw.reporting.common.model.*;
+import org.opentestsystem.rdw.reporting.common.model.AbstractAggregateReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.AggregateRow;
+import org.opentestsystem.rdw.reporting.common.model.District;
+import org.opentestsystem.rdw.reporting.common.model.Organization;
+import org.opentestsystem.rdw.reporting.common.model.ReportStatus;
+import org.opentestsystem.rdw.reporting.common.model.SchoolGradePrintableReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.StudentFilters;
 import org.opentestsystem.rdw.reporting.processor.model.UserReport;
 import org.opentestsystem.rdw.reporting.processor.web.ReportContentResource;
 import org.springframework.context.MessageSource;
@@ -33,7 +39,7 @@ import static org.mockito.Mockito.when;
 import static org.opentestsystem.rdw.reporting.common.util.MediaTypes.TEXT_CSV_UTF8;
 import static org.springframework.http.MediaType.APPLICATION_JSON;
 
-abstract class AbstractAggregateReportCsvHandlerWithFilteredSubgroupQueryTest<R extends AggregateRow, A extends AbstractSubgroupedAggregateReportQuery, B extends AbstractSubgroupedAggregateReportQuery.Builder<A, B>> extends AbstractAggregateReportCsvHandlerTest {
+abstract class AbstractAggregateReportCsvHandlerWithFilteredSubgroupQueryTest<R extends AggregateRow, A extends AbstractAggregateReportQuery, B extends AbstractAggregateReportQuery.Builder<A, B>> extends AbstractAggregateReportCsvHandlerTest {
 
     @Mock
     private MessageSource messageSource;
@@ -73,7 +79,7 @@ abstract class AbstractAggregateReportCsvHandlerWithFilteredSubgroupQueryTest<R 
                 .genderCodes(ImmutableSet.of("Female"))
                 .build());
 
-        final AbstractSubgroupedAggregateReportQuery query = getQueryBuilder()
+        final AbstractAggregateReportQuery query = getQueryBuilder()
                 .subgroups(subgroups)
                 .build();
