+++ report-processor/src/test/java/org/opentestsystem/rdw/reporting/processor/web/converter/ClaimAggregateReportCsvHandlerWithFilteredSubgroupQueryTest.java
@@ -5,7 +5,12 @@ import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import org.junit.runner.RunWith;
 import org.mockito.junit.MockitoJUnitRunner;
-import org.opentestsystem.rdw.reporting.common.model.*;
+import org.opentestsystem.rdw.reporting.common.model.ClaimAggregateReportQuery;
+import org.opentestsystem.rdw.reporting.common.model.ClaimRow;
+import org.opentestsystem.rdw.reporting.common.model.Dimension;
+import org.opentestsystem.rdw.reporting.common.model.DimensionType;
+import org.opentestsystem.rdw.reporting.common.model.Organization;
+import org.opentestsystem.rdw.reporting.common.model.StudentFilters;
 import org.springframework.context.MessageSource;
 
 import java.util.Map;
