+++ reporting-service/src/test/java/org/opentestsystem/rdw/reporting/exam/schoolgrade/JdbcSchoolGradeAssessmentRepositoryIT.java
@@ -10,7 +10,9 @@ import org.springframework.context.annotation.Import;
 import org.springframework.test.context.jdbc.Sql;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.opentestsystem.rdw.reporting.common.test.support.PermissionScopes.*;
+import static org.opentestsystem.rdw.reporting.common.test.support.PermissionScopes.individualOf;
+import static org.opentestsystem.rdw.reporting.common.test.support.PermissionScopes.permissions;
+import static org.opentestsystem.rdw.reporting.common.test.support.PermissionScopes.schools;
 
 @Import(JdbcSchoolGradeAssessmentRepository.class)
 @Sql(scripts = {"classpath:integration-test-data.sql"})
